Change Skeletal Keyboard Code to USB Example Project

+++ Skeletal_Keyboard_Code/driverlib/MSP430F5xx_6xx/adc12_a.h
@@ -1,3 +1,34 @@
+/* --COPYRIGHT--,BSD
+ * Copyright (c) 2016, Texas Instruments Incorporated
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions
+ * are met:
+ *
+ * *  Redistributions of source code must retain the above copyright
+ *    notice, this list of conditions and the following disclaimer.
+ *
+ * *  Redistributions in binary form must reproduce the above copyright
+ *    notice, this list of conditions and the following disclaimer in the
+ *    documentation and/or other materials provided with the distribution.
+ *
+ * *  Neither the name of Texas Instruments Incorporated nor the names of
+ *    its contributors may be used to endorse or promote products derived
+ *    from this software without specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+ * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO,
+ * THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR
+ * PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR
+ * CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
+ * EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
+ * PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS;
+ * OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY,
+ * WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR
+ * OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
+ * EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ * --/COPYRIGHT--*/
 //*****************************************************************************
 //
 // adc12_a.h - Driver for the ADC12_A Module.
@@ -29,7 +60,8 @@ extern "C"
 //! parameter.
 //
 //*****************************************************************************
-typedef struct ADC12_A_configureMemoryParam {
+typedef struct ADC12_A_configureMemoryParam
+{
     //! Is the selected memory buffer to set the configuration for.
     //! \n Valid values are:
     //! - \b ADC12_A_MEMORY_0 [Default]
@@ -92,7 +124,6 @@ typedef struct ADC12_A_configureMemoryParam {
     //!    end of the sequence.
     uint8_t endOfSequence;
 } ADC12_A_configureMemoryParam;
-
 
 //*****************************************************************************
 //
@@ -555,8 +586,7 @@ extern void ADC12_A_disableSamplingTimer(uint16_t baseAddress);
 //! positive reference voltage, the internal REF module must be used to control
 //! the voltage level. Note that if a conversion has been started with the
 //! startConversion() function, then a call to disableConversions() is required
-//! before this function may be called. If conversion is not disabled, this
-//! function does nothing.
+//! before this function may be called.
 //!
 //! \param baseAddress is the base address of the ADC12_A module.
 //! \param param is the pointer to struct for memory configuration.
@@ -714,8 +744,8 @@ extern void ADC12_A_clearInterrupt(uint16_t baseAddress,
 //! \return The current interrupt flag status for the corresponding mask.
 //
 //*****************************************************************************
-extern uint16_t ADC12_A_getInterruptStatus(uint16_t baseAddress,
-                                           uint16_t memoryInterruptFlagMask);
+extern uint8_t ADC12_A_getInterruptStatus(uint16_t baseAddress,
+                                          uint16_t memoryInterruptFlagMask);
 
 //*****************************************************************************
 //
@@ -1036,3 +1066,4 @@ extern uint16_t ADC12_A_isBusy(uint16_t baseAddress);
 
 #endif
 #endif // __MSP430WARE_ADC12_A_H__
+//Released_Version_5_20_06_02
